# modules/api/dashboard-api/src/main/resources/static/api-docs/openapi3.yaml
openapi: 3.0.1
info:
  title: Goodpon Dashboard API
  description: Goodpon Dashboard API Documentation
  version: 1.0.0
servers:
- url: http://localhost:8080
...
paths:
...
  /v1/merchants/{merchantId}/coupon-templates:
    post:
      tags:
      - Coupon Template
      summary: 쿠폰 템플릿 생성
      description: 쿠폰 템플릿 생성 API
      operationId: '쿠폰 템플릿 생성 - '
      parameters:
      - name: merchantId
        in: path
        description: 쿠폰 템플릿을 생성할 상점 ID
        required: true
        schema:
          type: string
      - name: Authorization
        in: header
        description: "인증 토큰 (Bearer {access-token})"
        required: true
        schema:
          type: string
        example: "Bearer {access-token}"
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCouponTemplateRequest'
            examples:
              쿠폰 템플릿 생성 - 실패 - 유효하지 않은 요청 데이터:
                value: "{\n  \"name\" : \"테스트 쿠폰\",\n  \"description\" : \"테스트 쿠폰\
                  \ 설명\",\n  \"minOrderAmount\" : 0,\n  \"discountType\" : \"PERCENTAGE\"\
                  ,\n  \"discountValue\" : 105,\n  \"maxDiscountAmount\" : 5000,\n\
                  \  \"issueStartDate\" : \"2025-07-18\",\n  \"issueEndDate\" : \"\
                  2025-07-16\",\n  \"validityDays\" : 3,\n  \"absoluteExpiryDate\"\
                  \ : \"2025-07-05\",\n  \"limitType\" : \"ISSUE_COUNT\",\n  \"maxIssueCount\"\
                  \ : 100,\n  \"maxRedeemCount\" : 200\n}"
              쿠폰 템플릿 생성 - 성공:
                value: "{\n  \"name\" : \"테스트 쿠폰\",\n  \"description\" : \"테스트 쿠폰\
                  \ 설명\",\n  \"minOrderAmount\" : 10000,\n  \"discountType\" : \"\
                  FIXED_AMOUNT\",\n  \"discountValue\" : 1000,\n  \"maxDiscountAmount\"\
                  \ : null,\n  \"issueStartDate\" : \"2025-07-13\",\n  \"issueEndDate\"\
                  \ : \"2025-07-20\",\n  \"validityDays\" : 30,\n  \"absoluteExpiryDate\"\
                  \ : \"2025-08-20\",\n  \"limitType\" : \"ISSUE_COUNT\",\n  \"maxIssueCount\"\
                  \ : 2000,\n  \"maxRedeemCount\" : null\n}"
              쿠폰 템플릿 생성 - 실패 - 존재하지 않는 상점:
                value: "{\n  \"name\" : \"테스트 쿠폰\",\n  \"description\" : \"테스트 쿠폰\
                  \ 설명\",\n  \"minOrderAmount\" : 10000,\n  \"discountType\" : \"\
                  FIXED_AMOUNT\",\n  \"discountValue\" : 1000,\n  \"maxDiscountAmount\"\
                  \ : null,\n  \"issueStartDate\" : \"2025-07-13\",\n  \"issueEndDate\"\
                  \ : \"2025-07-20\",\n  \"validityDays\" : 30,\n  \"absoluteExpiryDate\"\
                  \ : \"2025-08-20\",\n  \"limitType\" : \"ISSUE_COUNT\",\n  \"maxIssueCount\"\
                  \ : 2000,\n  \"maxRedeemCount\" : null\n}"
              쿠폰 템플릿 생성 - 실패 - 상점 접근 권한 없음:
                value: "{\n  \"name\" : \"테스트 쿠폰\",\n  \"description\" : \"테스트 쿠폰\
                  \ 설명\",\n  \"minOrderAmount\" : 10000,\n  \"discountType\" : \"\
                  FIXED_AMOUNT\",\n  \"discountValue\" : 1000,\n  \"maxDiscountAmount\"\
                  \ : null,\n  \"issueStartDate\" : \"2025-07-13\",\n  \"issueEndDate\"\
                  \ : \"2025-07-20\",\n  \"validityDays\" : 30,\n  \"absoluteExpiryDate\"\
                  \ : \"2025-08-20\",\n  \"limitType\" : \"ISSUE_COUNT\",\n  \"maxIssueCount\"\
                  \ : 2000,\n  \"maxRedeemCount\" : null\n}"
      responses:
        "400":
          description: "400"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/v1-merchants-merchantId-coupon-templates1838914843'
              examples:
                쿠폰 템플릿 생성 - 실패 - 유효하지 않은 요청 데이터:
                  value: "{\n  \"result\" : \"ERROR\",\n  \"data\" : null,\n  \"error\"\
                    \ : {\n    \"code\" : \"COUPON_TEMPLATE_VALIDATION_FAILED\",\n\
                    \    \"message\" : \"잘못된 쿠폰 템플릿 생성 요청입니다. 각 필드의 오류를 확인하세요.\",\n\
                    \    \"data\" : [ {\n      \"field\" : \"minOrderAmount\",\n \
                    \     \"message\" : \"쿠폰 사용 조건의 최소 주문 금액은 0보다 커야 합니다.\"\n    },\
                    \ {\n      \"field\" : \"discountValue\",\n      \"message\" :\
                    \ \"백분율 할인 값은 1~100 사이여야 합니다.\"\n    }, {\n      \"field\" : \"\
                    issueEndDate\",\n      \"message\" : \"쿠폰 발행 종료일은 발행 시작일보다 이전일\
                    \ 수 없습니다.\"\n    }, {\n      \"field\" : \"maxRedeemCount\",\n\
                    \      \"message\" : \"발행 제한 정책이 설정된 쿠폰은 사용 제한 수량을 함께 설정할 수 없습\
                    니다.\"\n    } ]\n  }\n}"
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse<CreateCouponTemplateResult>'
              examples:
                쿠폰 템플릿 생성 - 성공:
                  value: "{\n  \"result\" : \"SUCCESS\",\n  \"data\" : {\n    \"id\"\
                    \ : 1,\n    \"name\" : \"테스트 쿠폰\",\n    \"description\" : \"테스\
                    트 쿠폰 설명\",\n    \"merchantId\" : 1,\n    \"minOrderAmount\" :\
                    \ 10000,\n    \"discountType\" : \"FIXED_AMOUNT\",\n    \"discountValue\"\
                    \ : 1000,\n    \"maxDiscountAmount\" : null,\n    \"issueStartAt\"\
                    \ : \"2025-07-13T00:00:00\",\n    \"issueEndAt\" : \"2025-07-21T00:00:00\"\
                    ,\n    \"validityDays\" : 30,\n    \"absoluteExpiresAt\" : \"\
                    2025-08-21T00:00:00\",\n    \"limitType\" : \"ISSUE_COUNT\",\n\
                    \    \"maxIssueCount\" : 2000,\n    \"maxRedeemCount\" : null,\n\
                    \    \"status\" : \"DRAFT\"\n  },\n  \"error\" : null\n}"
        "404":
          description: "404"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/v1-merchants-merchantId1265661162'
              examples:
                쿠폰 템플릿 생성 - 실패 - 존재하지 않는 상점:
                  value: "{\n  \"result\" : \"ERROR\",\n  \"data\" : null,\n  \"error\"\
                    \ : {\n    \"code\" : \"MERCHANT_NOT_FOUND\",\n    \"message\"\
                    \ : \"존재하지 않는 상점입니다.\",\n    \"data\" : null\n  }\n}"
        "403":
          description: "403"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/v1-merchants-merchantId1265661162'
              examples:
                쿠폰 템플릿 생성 - 실패 - 상점 접근 권한 없음:
                  value: "{\n  \"result\" : \"ERROR\",\n  \"data\" : null,\n  \"error\"\
                    \ : {\n    \"code\" : \"NO_MERCHANT_ACCESS_PERMISSION\",\n   \
                    \ \"message\" : \"해당 상점에 접근할 수 있는 권한이 없습니다.\",\n    \"data\" :\
                    \ null\n  }\n}"
  /v1/merchants/{merchantId}/coupon-templates/{couponTemplateId}/publish:
    post:
      tags:
      - Coupon Template
      summary: 쿠폰 템플릿 발행
      description: 쿠폰 템플릿 발행 API
      operationId: '쿠폰 템플릿 발행 - '
      parameters:
      - name: merchantId
        in: path
        description: 쿠폰 템플릿이 있는 상점 ID
        required: true
        schema:
          type: string
      - name: couponTemplateId
        in: path
        description: 발행할 쿠폰 템플릿 ID
        required: true
        schema:
          type: string
      - name: Authorization
        in: header
        description: "인증 토큰 (Bearer {access-token})"
        required: true
        schema:
          type: string
        example: "Bearer {access-token}"
      responses:
        "400":
          description: "400"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/v1-merchants-merchantId1265661162'
              examples:
                쿠폰 템플릿 발행 - 실패 - 상점이 소유하지 않은 쿠폰 템플릿:
                  value: "{\n  \"result\" : \"ERROR\",\n  \"data\" : null,\n  \"error\"\
                    \ : {\n    \"code\" : \"COUPON_TEMPLATE_INVALID_STATUS_TO_PUBLISH\"\
                    ,\n    \"message\" : \"쿠폰 템플릿을 발행할 수 있는 상태가 아닙니다.\",\n    \"data\"\
                    \ : null\n  }\n}"
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse<PublishCouponTemplateResult>'
              examples:
                쿠폰 템플릿 발행 - 성공:
                  value: "{\n  \"result\" : \"SUCCESS\",\n  \"data\" : {\n    \"id\"\
                    \ : 1,\n    \"name\" : \"테스트 쿠폰\",\n    \"merchantId\" : 1,\n\
                    \    \"status\" : \"ISSUABLE\"\n  },\n  \"error\" : null\n}"
        "404":
          description: "404"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/v1-merchants-merchantId1265661162'
              examples:
                쿠폰 템플릿 발행 - 실패 - 존재하지 않는 쿠폰 템플릿:
                  value: "{\n  \"result\" : \"ERROR\",\n  \"data\" : null,\n  \"error\"\
                    \ : {\n    \"code\" : \"COUPON_TEMPLATE_NOT_FOUND\",\n    \"message\"\
                    \ : \"존재하지 않는 쿠폰 템플릿입니다.\",\n    \"data\" : null\n  }\n}"
                쿠폰 템플릿 발행 - 실패 - 존재하지 않는 상점:
                  value: "{\n  \"result\" : \"ERROR\",\n  \"data\" : null,\n  \"error\"\
                    \ : {\n    \"code\" : \"MERCHANT_NOT_FOUND\",\n    \"message\"\
                    \ : \"존재하지 않는 상점입니다.\",\n    \"data\" : null\n  }\n}"
        "403":
          description: "403"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/v1-merchants-merchantId1265661162'
              examples:
                쿠폰 템플릿 발행 - 실패 - 상점 접근 권한 없음:
                  value: "{\n  \"result\" : \"ERROR\",\n  \"data\" : null,\n  \"error\"\
                    \ : {\n    \"code\" : \"NO_MERCHANT_ACCESS_PERMISSION\",\n   \
                    \ \"message\" : \"해당 상점에 접근할 수 있는 권한이 없습니다.\",\n    \"data\" :\
                    \ null\n  }\n}"
components:
  schemas:
    v1-merchants-merchantId-coupon-templates1838914843:
      required:
      - data
      - error
      - result
      type: object
      properties:
        result:
          type: string
          description: 요청 결과 (SUCCESS/ERROR)
        error:
          required:
          - code
          - data
          - message
          type: object
          properties:
            code:
              type: string
              description: 오류 코드
            data:
              type: array
              description: 오류 세부 정보 (필드별 오류 메시지)
              items:
                required:
                - field
                - message
                type: object
                properties:
                  field:
                    type: string
                    description: "오류가 발생한 필드 이름 (minOrderAmount, discountValue, issueEndDate,\
                      \ maxIssueCount 등)"
                  message:
                    type: string
                    description: 해당 필드의 오류 메시지
            message:
              type: string
              description: 오류 메시지
          description: 오류 정보
...
    ApiResponse<CreateCouponTemplateResult>:
      title: ApiResponse<CreateCouponTemplateResult>
      required:
      - data
      - error
      - result
      type: object
      properties:
        result:
          type: string
          description: 요청 결과 (SUCCESS/ERROR)
        data:
          required:
          - description
          - discountType
          - discountValue
          - id
          - issueStartAt
          - limitType
          - merchantId
          - name
          - status
          type: object
          properties:
            maxIssueCount:
              type: number
              description: 최대 발급 수량 (ISSUE_COUNT 선택 시
              nullable: true
            issueEndAt:
              type: string
              description: 쿠폰 발급 종료 날짜
              nullable: true
            issueStartAt:
              type: string
              description: 쿠폰 발급 시작 날짜
            description:
              type: string
              description: 쿠폰 템플릿 설명
            absoluteExpiresAt:
              type: string
              description: 쿠폰 절대 만료 날짜
              nullable: true
            limitType:
              type: string
              description: "쿠폰 제한 유형 (ISSUE_COUNT, REDEEM_COUNT, NONE)"
            maxRedeemCount:
              type: number
              description: 최대 사용 수량 (REDEEM_COUNT 선택 시)
              nullable: true
            minOrderAmount:
              type: number
              description: 최소 주문 금액
              nullable: true
            merchantId:
              type: number
              description: 상점 ID
            maxDiscountAmount:
              type: number
              description: 최대 할인 금액 (PERCENTAGE 선택 시)
              nullable: true
            name:
              type: string
              description: 쿠폰 템플릿 이름
            discountType:
              type: string
              description: "할인 유형 (FIXED_AMOUNT, PERCENTAGE)"
            id:
              type: number
              description: 쿠폰 템플릿 ID
            discountValue:
              type: number
              description: 할인 금액 또는 비율
            status:
              type: string
              description: "쿠폰 템플릿 상태 (DRAFT, ACTIVE, EXPIRED 등)"
            validityDays:
              type: number
              description: 쿠폰 유효 기간 (일 단위)
              nullable: true
          description: 응답 데이터
...
    CreateCouponTemplateRequest:
      title: CreateCouponTemplateRequest
      required:
      - description
      - discountType
      - discountValue
      - issueStartDate
      - limitType
      - name
      type: object
      properties:
        maxIssueCount:
          type: number
          description: 최대 발급 수량 (ISSUE_COUNT 선택 시)
          nullable: true
        issueEndDate:
          type: string
          description: 쿠폰 발급 종료 날짜 (YYYY-MM-DD 형식)
          nullable: true
        description:
          type: string
          description: 쿠폰 템플릿 설명
        limitType:
          type: string
          description: "쿠폰 제한 유형 (ISSUE_COUNT, REDEEM_COUNT, NONE)"
        maxRedeemCount:
          type: number
          description: 최대 사용 수량 (REDEEM_COUNT 선택 시)
          nullable: true
        minOrderAmount:
          type: number
          description: 최소 주문 금액
          nullable: true
        issueStartDate:
          type: string
          description: 쿠폰 발급 시작 날짜 (YYYY-MM-DD 형식)
        maxDiscountAmount:
          type: number
          description: 최대 할인 금액 (PERCENTAGE 선택 시)
          nullable: true
        name:
          type: string
          description: 쿠폰 템플릿 이름
        discountType:
          type: string
          description: "할인 유형 (FIXED_AMOUNT, PERCENTAGE)"
        discountValue:
          type: number
          description: 할인 금액 또는 비율
        absoluteExpiryDate:
          type: string
          description: 쿠폰 절대 만료 날짜 (YYYY-MM-DD 형식)
          nullable: true
        validityDays:
          type: number
          description: 쿠폰 유효 기간 (일 단위)
          nullable: true
...
    ApiResponse<PublishCouponTemplateResult>:
      title: ApiResponse<PublishCouponTemplateResult>
      required:
      - data
      - error
      - result
      type: object
      properties:
        result:
          type: string
          description: 요청 결과 (SUCCESS/ERROR)
        data:
          required:
          - id
          - merchantId
          - name
          - status
          type: object
          properties:
            merchantId:
              type: number
              description: 상점 ID
            name:
              type: string
              description: 쿠폰 템플릿 이름
            id:
              type: number
              description: 쿠폰 템플릿 ID
            status:
              type: string
              description: 쿠폰 템플릿 상태 ISSUABLE
          description: 응답 데이터
...
    v1-merchants-merchantId1265661162:
      required:
      - data
      - error
      - result
      type: object
      properties:
        result:
          type: string
          description: 요청 결과 (SUCCESS/ERROR)
        error:
          required:
          - code
          - data
          - message
          type: object
          properties:
            code:
              type: string
              description: 오류 코드
            message:
              type: string
              description: 오류 메시지
          description: 오류 정보
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Authorization 헤더에 'Bearer {access-token}' 형식으로 전달되는 액세스 토큰"
security:
  - bearerAuth: []
# @formatter:on